Problem: A server file makes it into client bundle (#913)
Solution: Move the server part of the code to a separate file, as `usersStats.js` has to be loaded on the client as it's imported a few times on the client.
